+++ R/bmdboot.R
@@ -611,7 +611,7 @@ plot.bmdboot <- function(x, BMDtype = c("zSD", "xfold"), remove.infinite = TRUE,
   }
   if (BMD_log_transfo)
   {
-    g <- g + scale_x_log10()
+    g <- g + scale_x_log10() + xlab("BMD (in log scale)")
   }
   
   return(g)