+++ R/bmdplot.R
@@ -184,8 +184,7 @@ bmdplot <- function(extendedres, BMDtype = c("zSD", "xfold"),
   }
   
   if (BMD_log_transfo)
-    gg <- gg + scale_x_log10()
-  
+    gg <- gg + scale_x_log10() + xlab("BMD (in log scale)") else
   gg <- gg + xlab("BMD")
   
   if (!missing(shapeby))